Byteだと列の牛の数の合計値よりも小さくなる可能性があるのでIntegerに変更

+++ src/main/java/com/masudak/nimmt/core/Field.java
@@ -75,7 +75,7 @@ public class Field {
 	 */
 	public int getLineWithMinCows() {
 		int lineIndex = 0;
-		int cows = Byte.MAX_VALUE;
+		int cows = Integer.MAX_VALUE;
 		for (int i = 0; i < lines.size(); i++) {
 			int totalCows = lines.get(i).getTotalCows();
 			if (totalCows < cows) {
